Refactoring(server-info): reader moved subsystem-specific data to appropriate classes
