+++ netutils_linux_hardware/cpu.py
@@ -1,16 +1,13 @@
 # coding=utf-8
 
 from netutils_linux_hardware.grade import Grade
-from netutils_linux_hardware.parser import Parser
+from netutils_linux_hardware.parser import Parser, YAMLLike
 from netutils_linux_hardware.rater_math import extract
 from netutils_linux_hardware.subsystem import Subsystem
 
 
 class CPU(Subsystem):
     """ Everything about CPU: layouts, NUMA, HZ, L3, lscpu """
-
-    def parse(self):
-        pass
 
     def rate(self):
         cpuinfo = extract(self.data, ['cpu', 'info'])
@@ -25,6 +22,16 @@ class CPU(Subsystem):
                 'L3 cache': Grade.int(cpuinfo.get('L3 cache'), 1000, 30000),
                 'Vendor ID': Grade.str(cpuinfo.get('Vendor ID'), good=['GenuineIntel']),
             }, self.folding.SUBSYSTEM)
+
+    def parse(self):
+        output = {
+            'info': self.read(YAMLLike, 'lscpu_info'),
+            'layout': self.read(CPULayout, 'lscpu_layout'),
+        }
+        for key in ('CPU MHz', 'BogoMIPS'):
+            if output.get('info', {}).get(key):
+                output['info'][key] = int(output['info'][key])
+        return output
 
 
 class CPULayout(Parser):
