+++ netutils_linux_hardware/subsystem.py
@@ -1,5 +1,6 @@
 # coding=utf-8
 
+import os
 from abc import abstractmethod
 
 from netutils_linux_hardware.folding import Folding
@@ -8,18 +9,14 @@ from netutils_linux_hardware.folding import Folding
 class Subsystem(object):
     """ Base class for CPU, Memory, etc """
 
-    def __init__(self, data=None, folding=None):
+    def __init__(self, data=None, folding=None, datadir=None):
+        self.datadir = datadir
         self.data = data
         self.folding = folding
 
     @abstractmethod
     def collect(self):
         """ Collecting required data from host """
-        pass
-
-    @abstractmethod
-    def read(self):
-        """ Read saved data from datadir """
         pass
 
     @abstractmethod
@@ -35,3 +32,11 @@ class Subsystem(object):
     def map(self, func, key):
         items = self.data.get(key)
         return self.folding.fold(dict((item, func(item)) for item in items), Folding.SUBSYSTEM) if items else 1
+
+    def path(self, filename):
+        assert self.datadir
+        assert filename
+        return os.path.join(self.datadir, filename)
+
+    def read(self, func, filename):
+        return func(self.path(filename)).result
