+++ netutils_linux_hardware/yaml.py
@@ -1,0 +1,8 @@
+# coding=utf-8
+
+import yaml
+
+
+def dict2yaml(d):
+    """ Converts nested dicts in pretty YAML """
+    return yaml.dump(d, default_flow_style=False).strip()
